add people interface
