+++ main.rb
@@ -2,11 +2,12 @@ require './book'
 require './student'
 require './teacher'
 require './rental'
+require './people_manager'
 class App
   def initialize
     @list_books = []
-    @list_people = []
     @list_rentals = []
+    @people = PeopleManager.new
   end
 
   # Books Block
@@ -40,73 +41,73 @@ class App
 
   #============================
   # People Block
-  def insert_student
-    print 'Insert Name:'
-    name = gets.chomp
-    print 'Insert Age:'
-    age = gets.chomp
-    print 'Has permission? [y, n] '
-    permission = gets.chomp
-    permission = permission == 'y'
-    @list_people.push(Student.new(age: age, name: name, parent_permission: permission))
-    puts 'Student created Succesfully'
-    show_menu
-  end
+  # def insert_student
+  #   print 'Insert Name:'
+  #   name = gets.chomp
+  #   print 'Insert Age:'
+  #   age = gets.chomp
+  #   print 'Has permission? [y, n] '
+  #   permission = gets.chomp
+  #   permission = permission == 'y'
+  #   @list_people.push(Student.new(age: age, name: name, parent_permission: permission))
+  #   puts 'Student created Succesfully'
+  #   show_menu
+  # end
 
-  def insert_teacher
-    print 'Insert Name: '
-    name = gets.chomp
-    print 'Insert Age: '
-    age = gets.chomp
-    print 'Insert Specialization: '
-    spec = gets.chomp
-    parent_auto = true
-    @list_people.push(Teacher.new(age: age, name: name, specialization: spec, parent_permission: parent_auto))
-    puts 'Teacher created Succesfully'
-    show_menu
-  end
+  # def insert_teacher
+  #   print 'Insert Name: '
+  #   name = gets.chomp
+  #   print 'Insert Age: '
+  #   age = gets.chomp
+  #   print 'Insert Specialization: '
+  #   spec = gets.chomp
+  #   parent_auto = true
+  #   @list_people.push(Teacher.new(age: age, name: name, specialization: spec, parent_permission: parent_auto))
+  #   puts 'Teacher created Succesfully'
+  #   show_menu
+  # end
 
-  def insert_people
-    puts 'which kind of person want to registered ? [1: Student 2: Teacher 3: Cancel]'
-    option = gets.chomp
-    case option
-    when '1'
-      insert_student
-    when '2'
-      insert_teacher
-    else
-      show_menu
-    end
-  end
+  # def insert_people
+  #   puts 'which kind of person want to registered ? [1: Student 2: Teacher 3: Cancel]'
+  #   option = gets.chomp
+  #   case option
+  #   when '1'
+  #     insert_student
+  #   when '2'
+  #     insert_teacher
+  #   else
+  #     show_menu
+  #   end
+  # end
 
-  def show_people
-    if @list_people.any? == false
-      puts "There's no people registered. Press 1 to return"
-      selected = gets.chomp
-      case selected
-      when '1'
-        show_menu
-      end
-    else
-      display_people
-    end
-  end
+  # def show_people
+  #   if @list_people.any? == false
+  #     puts "There's no people registered. Press 1 to return"
+  #     selected = gets.chomp
+  #     case selected
+  #     when '1'
+  #       show_menu
+  #     end
+  #   else
+  #     display_people
+  #   end
+  # end
 
-  def display_people
-    @list_people.each do |b|
-      if b.is_a?(Student)
-        puts " [Student] Name: #{b.name} ID: #{b.id} Age: #{b.age}"
-      else
-        puts " [Teacher] Name: #{b.name} ID: #{b.id} Age: #{b.age}"
-      end
-    end
-    puts 'Press 1 to return'
-    option = gets.chomp
-    case option
-    when '1'
-      show_menu
-    end
-  end
+  # def display_people
+  #   @list_people.each do |b|
+  #     if b.is_a?(Student)
+  #       puts " [Student] Name: #{b.name} ID: #{b.id} Age: #{b.age}"
+  #     else
+  #       puts " [Teacher] Name: #{b.name} ID: #{b.id} Age: #{b.age}"
+  #     end
+  #   end
+  #   puts 'Press 1 to return'
+  #   option = gets.chomp
+  #   case option
+  #   when '1'
+  #     show_menu
+  #   end
+  # end
 
   #============================
   # Rentals Block
@@ -144,9 +145,9 @@ class App
     when '1'
       show_books
     when '2'
-      show_people
+      @people.show_people
     when '3'
-      insert_people
+      @people.insert_people
     when '4'
       insert_book
     when '5'
@@ -160,11 +161,14 @@ class App
   end
 
   def show_menu
-    puts "\n\nHello, Please choose an option selecting a number:"
-    puts "1) List all Books\n2) List all People\n3) Create a Person\n4) Create a Book\n"
-    puts "5) Create a Rental\n6) List all rentals for a given person id\n7) Exit"
-    option = gets.chomp
-    read_input(option)
+    loop do
+      puts "\n\nHello, Please choose an option selecting a number:"
+      puts "1) List all Books\n2) List all People\n3) Create a Person\n4) Create a Book\n"
+      puts "5) Create a Rental\n6) List all rentals for a given person id\n7) Exit"
+      option = gets.chomp
+      break if option == '7'
+      read_input(option)
+    end
   end
 end
